Delete some more properties from created snapshots.
Some of these additional properties are VIO-specific.

In addition, since we are removing properties, it's not an error if
they do not exist; but `openstack image unset` will still fail, so
ignore errors.  (This results in incorrect behavior in case a property
exists but cannot be deleted; but I expect that in this case things
will have gone south well before this point...)