CentOS/RHEL: Install package `redhat-lsb-core`.
We used to ensure that `lsb_release` is installed by running `yum
install /usr/bin/lsb_release`.  While this still works from the
command-line, it fails in Ansible with this error:

    TASK [Ensure `/usr/bin/lsb_release` is installed] ***
    fatal: [172.31.116.4]: FAILED! => {"changed": false, "msg": "No group usr/bin/lsb_release available.", "results": []}

I guess that Ansible changed the syntax of the `yum:` action so that a
name containing slashes is considered a group name -- but it's faster
to just install the right package name than to investigate this issue
to the bottom.